+++ node/resolvers/product/upload.ts
@@ -1,12 +1,9 @@
 import { ReadStream } from 'fs'
 
 import {
-  BUCKET_NAME,
-  calculateExportProcessTime,
+  PRODUCT_BUCKET,
   PRODUCT_TRANSLATION_UPLOAD,
-  CALLS_PER_MINUTE,
-  parseStreamToJSON,
-  uploadEntriesAsync,
+  uploadTranslations,
 } from '../../utils'
 
 const uploadProductTranslations = async (
@@ -14,64 +11,14 @@ const uploadProductTranslations = async (
   { products, locale }: { products: UploadFile<ReadStream>; locale: string },
   ctx: Context
 ) => {
-  const {
-    clients: { catalogGQL, licenseManager, vbase },
-    vtex: { adminUserAuthToken, requestId },
-  } = ctx
-
-  const { createReadStream } = await products
-
-  const productStream = createReadStream()
-
-  const productsParsed = await parseStreamToJSON<ProductTranslationInput>(
-    productStream
-  )
-
-  const {
-    profile: { email },
-  } = await licenseManager.getTopbarData(adminUserAuthToken as string)
-
-  const allTranslationsMade = await vbase.getJSON<string[]>(
-    BUCKET_NAME,
-    PRODUCT_TRANSLATION_UPLOAD,
-    true
-  )
-
-  const updateRequests = allTranslationsMade
-    ? [requestId, ...allTranslationsMade]
-    : [requestId]
-
-  await vbase.saveJSON<string[]>(
-    BUCKET_NAME,
-    PRODUCT_TRANSLATION_UPLOAD,
-    updateRequests
-  )
-
-  const requestInfo = {
-    requestId,
+  const params: UploadTranslations<ProductTranslationInput> = {
+    entries: products,
     locale,
-    translatedBy: email,
-    createdAt: new Date(),
-    estimatedTime: calculateExportProcessTime(
-      productsParsed.length,
-      CALLS_PER_MINUTE
-    ),
+    bucket: PRODUCT_BUCKET,
+    path: PRODUCT_TRANSLATION_UPLOAD,
+    translateEntry: ctx?.clients?.catalogGQL?.translateProduct,
   }
-
-  await vbase.saveJSON<UploadRequest>(BUCKET_NAME, requestId, requestInfo)
-
-  uploadEntriesAsync<ProductTranslationInput>(
-    {
-      entries: productsParsed,
-      requestId,
-      locale,
-      bucket: BUCKET_NAME,
-      translateEntry: catalogGQL?.translateProduct,
-    },
-    { vbase }
-  )
-
-  return requestId
+  return uploadTranslations(params, ctx)
 }
 
 const productTranslationsUploadRequests = async (
@@ -80,7 +27,7 @@ const productTranslationsUploadRequests = async (
   ctx: Context
 ) =>
   ctx.clients.vbase.getJSON<string[]>(
-    BUCKET_NAME,
+    PRODUCT_BUCKET,
     PRODUCT_TRANSLATION_UPLOAD,
     true
   )