+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Bulk import and export for `Collections`
+- Import all `Categories` translations in batch
+- Apply `ImportEntriesModal` for import translations in `Brand` & `Product` process
 
 ## [1.3.1] - 2021-12-14
 ### Added